+++ crates/rest_macro_core/src/compiler/codegen.rs
@@ -430,6 +430,9 @@ use self::service_tokens::*;
 
 
 
+mod garde_tokens;
+use self::garde_tokens::*;
+
 mod resource_struct_tokens;
 use self::resource_struct_tokens::*;
